renamed title component
