+++ src/app/components/SectionTitle.tsx
